+++ src/main/java/io/github/thebusybiscuit/slimefun4/api/network/NetworkComponent.java
@@ -1,5 +1,7 @@
 package io.github.thebusybiscuit.slimefun4.api.network;
 
+import io.github.thebusybiscuit.slimefun4.core.networks.NetworkManager;
+
 /**
  * This enum holds the different types of components a {@link Network} can have.
  * It is used for classification of nodes inside the {@link Network}.
