+++ src/main/java/io/github/thebusybiscuit/slimefun4/core/networks/package-info.java
@@ -1,0 +1,6 @@
+/**
+ * This package provides the core functionality for the {@link io.github.thebusybiscuit.slimefun4.api.network.Network}
+ * class, such as the {@link io.github.thebusybiscuit.slimefun4.core.networks.NetworkManager} and also the subpackages
+ * for actual implementations of the {@link io.github.thebusybiscuit.slimefun4.api.network.Network} class.
+ */
+package io.github.thebusybiscuit.slimefun4.core.network;
