+++ src/main/java/io/github/thebusybiscuit/slimefun4/implementation/listeners/NetworkListener.java
@@ -7,7 +7,7 @@ import org.bukkit.event.block.BlockBreakEvent;
 import org.bukkit.event.block.BlockPlaceEvent;
 
 import io.github.thebusybiscuit.slimefun4.api.network.Network;
-import io.github.thebusybiscuit.slimefun4.api.network.NetworkManager;
+import io.github.thebusybiscuit.slimefun4.core.networks.NetworkManager;
 import me.mrCookieSlime.Slimefun.SlimefunPlugin;
 
 /**
